move io tests
