+++ test/integration/io/npm-registry.test.ts
@@ -1,7 +1,7 @@
 import { userInfo } from "os";
 import { Registry } from "../../../src/domain/registry";
 import { makeNpmFetchOptions } from "../../../src/io/npm-registry";
-import { exampleRegistryUrl } from "../domain/data-registry";
+import { exampleRegistryUrl } from "../../unit/domain/data-registry";
 import { Options as FetchOptions } from "npm-registry-fetch";
 import { NpmAuth } from "another-npm-registry-client";
