+++ test/integration/io/registry-client.mock.ts
