+++ test/integration/io/special-paths.test.ts
